+++ example/hello-mvc/webapp/listView.rsp.dart
@@ -7,7 +7,8 @@ Future listView(HttpConnect connect, {String path, List<FileInfo> infos}) { //#3
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   response.write("""<!DOCTYPE html>
 <html>
@@ -58,5 +59,5 @@ Future listView(HttpConnect connect, {String path, List<FileInfo> infos}) { //#3
 </html>
 """); //#23
 
-  return Rsp.nnf();
+  return new Future.value();
 }