+++ example/hello-rsp/webapp/helloView.rsp.dart
@@ -7,7 +7,8 @@ Future helloView(HttpConnect connect) { //#3
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   response.write("""<!DOCTYPE html>
 <html>
@@ -35,5 +36,5 @@ Future helloView(HttpConnect connect) { //#3
 </html>
 """); //#12
 
-  return Rsp.nnf();
+  return new Future.value();
 }