+++ example/hello-templating/webapp/classic.rsp.dart
@@ -7,7 +7,8 @@ Future classic(HttpConnect connect, {header, sidebar, body, footer}) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   response.write("""<div>
   <div class="header">
@@ -49,5 +50,5 @@ Future classic(HttpConnect connect, {header, sidebar, body, footer}) { //#2
 </div>
 """); //#13
 
-  return Rsp.nnf();
+  return new Future.value();
 }