+++ example/hello-templating/webapp/home.rsp.dart
@@ -7,7 +7,8 @@ Future home(HttpConnect connect) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   response.write("""<!DOCTYPE html>
 <html>
@@ -67,7 +68,7 @@ Future home(HttpConnect connect) { //#2
 </html>
 """); //#28
 
-          return Rsp.nnf();
+          return new Future.value();
         }); //end-of-include
       }); //end-of-include
     }); //end-of-include