+++ example/hello-templating/webapp/sidebar.rsp.dart
@@ -7,7 +7,8 @@ Future sidebar(HttpConnect connect) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   response.write("""<ul>
   <li>This is a sidebar.</li>
@@ -20,5 +21,5 @@ Future sidebar(HttpConnect connect) { //#2
 </ul>
 """); //#4
 
-  return Rsp.nnf();
+  return new Future.value();
 }