+++ test/features/webapp/forwarderView.rsp.dart
@@ -7,12 +7,13 @@ Future forwarderView(HttpConnect connect) { //#3
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   if (true) { //if#3
 
     return connect.forward("/forward"); //forward#4
   } //if
 
-  return Rsp.nnf();
+  return new Future.value();
 }