+++ test/features/webapp/fragView.rsp.dart
@@ -7,7 +7,8 @@ Future fragView(HttpConnect connect, {Map infos: const {}, header, footer}) { //
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   if (header != null) { //if#2
 
@@ -72,5 +73,5 @@ Future fragView(HttpConnect connect, {Map infos: const {}, header, footer}) { //
 """); //#19
   } //if
 
-  return Rsp.nnf();
+  return new Future.value();
 }