+++ test/features/webapp/includerView.rsp.dart
@@ -7,7 +7,8 @@ Future includerView(HttpConnect connect) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 final infos = {
   "fruits": ["apple", "orange", "lemon"],
   "cars": ["bmw", "audi", "honda"]
@@ -71,7 +72,7 @@ final infos = {
 </html>
 """); //#35
 
-          return Rsp.nnf();
+          return new Future.value();
         }); //end-of-include
       }); //end-of-include
     }); //end-of-include