+++ test/features/webapp/json.rsp.dart
@@ -7,7 +7,8 @@ Future json(HttpConnect connect) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 var map = {
   "first": [123, "abc"],
   "second": true
@@ -37,5 +38,5 @@ document.getElementById("show").innerHTML = out;
 </html>
 """); //#14
 
-  return Rsp.nnf();
+  return new Future.value();
 }