+++ test/features/webapp/searchResult.rsp.dart
@@ -7,7 +7,8 @@ Future searchResult(HttpConnect connect, {criteria}) { //#2
   var _t0_, _cs_ = new List<HttpConnect>();
   HttpRequest request = connect.request;
   HttpResponse response = connect.response;
-  Rsp.init(connect, "text/html; charset=utf-8");
+  if (!Rsp.init(connect, "text/html; charset=utf-8"))
+    return new Future.value();
 
   response.write("""<html>
   <head>
@@ -47,5 +48,5 @@ Future searchResult(HttpConnect connect, {criteria}) { //#2
 </html>
 """); //#14
 
-  return Rsp.nnf();
+  return new Future.value();
 }